requirements.txt
albumentations==1.3.1
clip==1.0
dlib==19.24.2
easydict==1.11
efficientnet_pytorch==0.7.1
einops==0.8.0
matplotlib==3.5.3
nltk==3.9.1
numpy==1.21.6
omegaconf==2.3.0
open_clip_torch==2.0.2
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
pandas==1.3.5
Pillow==9.5.0
Pillow==10.4.0
pytorch_lightning==1.5.0
safetensors==0.4.5
schedule==1.2.1
scikit_image==0.19.3
scikit_learn==1.0.2
scipy==1.7.3
tensorboardX==2.6.2.2
tensorboardX==2.6.2.2
timm==0.4.12
torch==1.13.1
torchvision==0.14.1
tqdm==4.66.1
transform==1.4.3
transformers==4.30.2
xformers==0.0.28.post1
